strip timestamp from log lines on Azure and GHA
Both Azure and GHA add the timestamp of each line along with the line
itself, separated by a space. This was messing up "log variabiles"
detection, as the detector tried to find a line starting with [ while
all lines started with the timestamp.

Since this behavior is specific to Azure and GHA I added a new method to
the CiProvider trait, which filters out timestamps from log lines.  The
method is currently implemented by all providers, but that might not
hold true anymore if we migrate to yet another provider in the future.