handle CI providers returning merge commits vs head commits
When dealing with PRs, both Azure Pipelines and Travis CI returned the
SHA of the merge commit as the commit being tested during the build,
while when dealing with branches they returned the head of the branch.

GitHub Actions behaves differently, as it always returns the head of the
branch, even in PRs.

To address the problem, this encodes in the type system whether a commit
is a merge commit or a head commit. Then, if the server receives a merge
commit it will fetch the related head commit with the GitHub API.