server: avoid skipping builds if we can't fetch them
Due to GitHub API inconsistencies, it might happen that fetching the
build from the API the first time fails. This will return an error,
causing a return before the ID could be removed from the "seen" list.

This fixes the problem by moving adding a crate to the "seen" list after
such requests.